+++ src/commands/syncto.ts
@@ -36,19 +36,23 @@ import Users from '../classes/users'
  * 
  */
 export default class Syncto extends Command {
+
+    verbose = false
+
+    echo = {}
+
     luksName = 'luks-eggs-backup'
+
     luksFile = `/run/live/medium/live/${this.luksName}`
+
     luksDevice = `/dev/mapper/${this.luksName}`
+
     luksMountpoint = '/tmp/eggs-backup'
 
     static description = 'Backup users, server and datas to luks-eggs-backup'
 
     static flags = {
         delete: Flags.string({ description: 'rsync --delete delete extraneous files from dest dirs' }),
-        exclude: Flags.string({ description: 'same as rsync --exclude=PATTERN exclude files matching PATTERN'}),
-        excludeFrom: Flags.string({ description: 'same as rsync --exclude-from=FILE read exclude patterns from FILE' }),
-        include: Flags.string({ description: 'same as rsync --include=PATTERN include files matching PATTERN'}),
-        includeFrom: Flags.string({ description: 'same as rsync --include-from=FILE read include patterns from FILE' }),
         file: Flags.string({ char: 'f', description: "file LUKS volume encrypted" }),
         help: Flags.help({ char: 'h' }),
         verbose: Flags.boolean({ char: 'v', description: 'verbose' })
@@ -64,17 +68,22 @@ export default class Syncto extends Command {
     async run(): Promise<void> {
 
         const { flags } = await this.parse(Syncto)
-        let verbose = false
+
         if (flags.verbose) {
-            verbose = true
+            this.verbose = true
         }
+        this.echo = Utils.setEcho(this.verbose)
 
         let fileVolume = ''
         if (flags.file) {
             fileVolume = flags.file
         }
 
-        const echo = Utils.setEcho(verbose)
+        let destDelete = false
+        if (flags.delete) {
+            destDelete = true
+        }
+
         if (Utils.isRoot(this.id)) {
             /**
              * restore con file
@@ -88,27 +97,24 @@ export default class Syncto extends Command {
             this.luksDevice = `/dev/mapper/${this.luksName}`
             this.luksMountpoint = '/tmp/eggs-backup'
             if (!fs.existsSync(fileVolume)) {
-                await this.luksCreate(verbose)
+                await this.luksCreate()
             } else {
                 Utils.warning(`LUKS volume: ${this.luksFile} exist, don't need create`)
             }
             if (fs.existsSync(fileVolume)) {
-                await this.luksOpen(verbose)
-                await this.backup(verbose)
-                await this.luksClose(verbose)
+                await this.luksOpen()
+                await this.backup(destDelete)
+                await this.luksClose()
             }
         }
     }
 
 
-
     /**
      *
-     * @param verbose
      */
-    async backup(verbose = false) {
-        const echo = Utils.setEcho(verbose)
-        if (verbose) {
+    async backup(destDelete = false) {
+        if (this.verbose) {
             Utils.warning('backup')
         }
         const usersArray = await this.usersFill()
@@ -116,18 +122,22 @@ export default class Syncto extends Command {
         for (let i = 0; i < usersArray.length; i++) {
             if (usersArray[i].saveIt) {
                 if (fs.existsSync(usersArray[i].home)) {
-                    await exec(`mkdir -p ${this.luksMountpoint}/ROOT${usersArray[i].home}`, echo)
+                    await exec(`mkdir -p ${this.luksMountpoint}/ROOT${usersArray[i].home}`, this.echo)
                     const source = usersArray[i].home
                     let dest = this.luksMountpoint + '/ROOT' + usersArray[i].home
                     dest = dest.substring(0, dest.lastIndexOf('/'))
-                    await exec(`rsync --archive ${source} ${dest}`, echo)
+                    let cmd = `rsync --archive ${source} ${dest}`
+                    if (destDelete) {
+                        cmd = `rsync --archive --delete ${source} ${dest}`
+                    }
+                    await exec(cmd, this.echo)
                 }
             }
         }
-        await exec(`mkdir -p ${this.luksMountpoint}/etc`, echo)
-        await exec(`cp /etc/passwd ${this.luksMountpoint}/etc`, echo)
-        await exec(`cp /etc/shadow ${this.luksMountpoint}/etc`, echo)
-        await exec(`cp /etc/group ${this.luksMountpoint}/etc`, echo)
+        await exec(`mkdir -p ${this.luksMountpoint}/etc`, this.echo)
+        await exec(`cp /etc/passwd ${this.luksMountpoint}/etc`, this.echo)
+        await exec(`cp /etc/shadow ${this.luksMountpoint}/etc`, this.echo)
+        await exec(`cp /etc/group ${this.luksMountpoint}/etc`, this.echo)
     }
 
 
@@ -153,10 +163,7 @@ export default class Syncto extends Command {
     /**
      * 
      */
-    async luksCreate(verbose = false) {
-        const echo = Utils.setEcho(verbose)
-        const echoYes = Utils.setEcho(true)
-
+    async luksCreate() {
         Utils.warning(`Creating LUKS Volume on ${this.luksFile}`)
 
         let totalSize = 0
@@ -191,34 +198,31 @@ export default class Syncto extends Command {
         }
 
         Utils.warning(`Creating an encrypted file ${this.luksFile} blocks=${blocks}, block size: ${blockSize}, size: ${Utils.formatBytes(blocks * blockSize)}`)
-        await exec(`dd if=/dev/zero of=${this.luksFile} bs=${blockSize} count=${blocks}`, echo)
+        await exec(`dd if=/dev/zero of=${this.luksFile} bs=${blockSize} count=${blocks}`, this.echo)
 
         // find first unused device
-        let findFirstUnusedDevice = await exec(`losetup -f`, { echo: verbose, ignore: false, capture: true })
+        let findFirstUnusedDevice = await exec(`losetup -f`, { echo: this.verbose, ignore: false, capture: true })
         let firstUnusedDevice = ''
         if (findFirstUnusedDevice.code !== 0) {
-            Utils.warning(`Error: ${findFirstUnusedDevice.code} ${findFirstUnusedDevice.data}`)
-            process.exit(1)
+            Utils.pressKeyToExit(`Error: ${findFirstUnusedDevice.code} ${findFirstUnusedDevice.data}`)
         } else {
             firstUnusedDevice = findFirstUnusedDevice.data.trim()
         }
-        await exec(`losetup ${firstUnusedDevice} ${this.luksFile}`, echo)
+        await exec(`losetup ${firstUnusedDevice} ${this.luksFile}`, this.echo)
 
         Utils.warning('Enter a large string of random text below to setup the pre-encryption')
-        await exec(`cryptsetup -y -v --type luks2 luksFormat  ${this.luksFile}`, echoYes)
+        await exec(`cryptsetup -y -v --type luks2 luksFormat  ${this.luksFile}`, Utils.setEcho(true))
 
         Utils.warning(`Enter the desired passphrase for the encrypted ${this.luksName} below`)
-        let crytoSetup = await exec(`cryptsetup luksOpen --type luks2 ${this.luksFile} ${this.luksName}`, echoYes)
+        let crytoSetup = await exec(`cryptsetup luksOpen --type luks2 ${this.luksFile} ${this.luksName}`, Utils.setEcho(true))
         if (crytoSetup.code !== 0) {
-            Utils.warning(`Error: ${crytoSetup.code} ${crytoSetup.data}`)
-            process.exit(1)
+            Utils.pressKeyToExit(`Error: ${crytoSetup.code} ${crytoSetup.data}`)
         }
 
         Utils.warning(`Formatting ${this.luksDevice} to ext2`)
-        let formatting = await exec(`sudo mkfs.ext2 ${this.luksDevice}`, echo)
+        let formatting = await exec(`sudo mkfs.ext2 ${this.luksDevice}`, this.echo)
         if (formatting.code !== 0) {
-            Utils.warning(`Error: ${formatting.code} ${formatting.data}`)
-            process.exit(1)
+            Utils.pressKeyToExit(`Error: ${formatting.code} ${formatting.data}`)
         }
         // this.luksClose()
     }
@@ -227,24 +231,21 @@ export default class Syncto extends Command {
     /**
      * 
      */
-    async luksOpen(verbose = false) {
-        const echo = Utils.setEcho(verbose)
-        const echoYes = Utils.setEcho(true) // echoYes serve solo per cryptsetup luksOpen
-
+    async luksOpen() {
         if (!fs.existsSync(this.luksDevice)) {
             Utils.warning(`LUKS open volume: ${this.luksName}`)
-            await exec(`cryptsetup luksOpen --type luks2 ${this.luksFile} ${this.luksName}`, echoYes)
+            await exec(`cryptsetup luksOpen --type luks2 ${this.luksFile} ${this.luksName}`, Utils.setEcho(true))
         } else {
             Utils.warning(`LUKS volume: ${this.luksName} already open`)
         }
 
         if (!fs.existsSync(this.luksMountpoint)) {
-            await exec(`mkdir -p ${this.luksMountpoint}`, echo)
+            await exec(`mkdir -p ${this.luksMountpoint}`, this.echo)
         }
 
         if (!Utils.isMountpoint(this.luksMountpoint)) {
             Utils.warning(`mount volume: ${this.luksDevice} on ${this.luksMountpoint}`)
-            await exec(`mount ${this.luksDevice} ${this.luksMountpoint}`, echo)
+            await exec(`mount ${this.luksDevice} ${this.luksMountpoint}`, this.echo)
         } else {
             Utils.warning(`mount volume: ${this.luksDevice} already mounted on ${this.luksMountpoint}`)
         }
@@ -254,16 +255,14 @@ export default class Syncto extends Command {
     /**
     * 
     */
-    async luksClose(verbose = false) {
-        const echo = Utils.setEcho(verbose)
-
+    async luksClose() {
         if (Utils.isMountpoint(this.luksMountpoint)) {
-            await exec(`umount ${this.luksMountpoint}`, echo)
+            await exec(`umount ${this.luksMountpoint}`, this.echo)
         }
 
         if (fs.existsSync(this.luksDevice)) {
             Utils.warning(`LUKS close volume: ${this.luksName}`)
-            await exec(`cryptsetup luksClose ${this.luksName}`, echo)
+            await exec(`cryptsetup luksClose ${this.luksName}`, this.echo)
         }
     }